[Feat] Spacemacs Lean! Total Rewrite, Very Light.
Switched to Spacemacs Base which made everything faster
through a lot of crap out of the init, only what is
totally necessary was left (there was a lot of
garbage to throw out)
This certainly feels a lot faster.
Removing Spaceline helped a lot (profile-report) showed
it was causing many problesm

Switching to base made most of the difference.