Enabled PDF layer and set up docview autoload
This involved using require which is ordinarily not recommended
in /doom/ emacs, but it isn't loaded otherwise so I have to do it.

I also set the default app as pdfview for org-mode so if I do want
to open the pdf file externally I will need to configure a function
to call `xdg-open *.pdf`.